opt: performance improvements for histograms
This commit improves the performance of histograms in the optimizer
by avoiding allocating and copying the full histogram unless strictly
necessary. Additionally, it changes the code for filtering histograms
to use binary search instead of performing a linear scan.

Release note: None

+++ pkg/sql/opt/testutils/testcat/test_catalog.go
@@ -984,8 +984,8 @@ func (ts *TableStat) Histogram() []cat.HistogramBucket {
 			panic(err)
 		}
 		histogram[i] = cat.HistogramBucket{
-			NumEq:         uint64(bucket.NumEq),
-			NumRange:      uint64(bucket.NumRange),
+			NumEq:         float64(bucket.NumEq),
+			NumRange:      float64(bucket.NumRange),
 			DistinctRange: bucket.DistinctRange,
 			UpperBound:    datum,
 		}
